fix: Subcomponent config on instance/version
When determining which subcomponents to generate, permit the definition
of the subcomponents on the instance or version configuration.

Previously, subcomponents could only be defined at the component type
level.

The use case for this was an API lambda function when one instance was
based on python and another on javascript. Without this fix, it was only
possible to have one runtime defined for the entire lambda component.