refactor: Namespace matching support
Place the code that does namespace matching into its own routine and use
this routine when processing occurrence settings.

This is to enable reuse of this functionality as part of the dynamic
CMDB work.

Essentially the plan is to use the same logic for assembly of solutions,
potentially allowing the namespace concept we use for settings to
enable selective inclusion of solution fragments based on name not
hierarchical position.