feat: deployment mode and group refernce
Adds two new reference objects

- Deployment mode: provides a way to sequence deployments based on
deployment groups. A deployment mode specifies which groups it will
deploy and this can be provided through 2 methods.
   - Priority: this is a dynamic approach which uses the priroity of a
deployment group combined with a pattern match on the name of the group.
Groups are ordered based on their priority with the option to reverse
the priority if required
   - Explicit: is a named ordered array of deployment groups.

- Deployment Group: represents a group of deployment units. Groups do
not specify thier membership. Instead units nominate a group to be a
member of. Deployment Groups also provide support for level filtering
and resource sets which allow you to create resource type based
deployment units