feat: component deployment attributes
Creates a specifc namespace in component attributes to handle the
attributes required to manage the deployment of an occurrence
The attributes required for a deployable occurrence are the same however
the default values need to be different to provides compatability and
usability. I.e. the default deployment group for the basline component
should be different from the apigateway component.

This adds a new macro to components addComponentDeployment which should
be invoked for components or subcomponents which support being deployed
in their own right. For sub components which wouldn't work with being
deployed on their own this shouldn't be used.

For the baseline component the option to lock the defaults is also
avaialble. This allows the component developer to only allow the
defaults to only be the values available.

This also deprecates the existing DeploymentUnits global attribute in
favour of te namespaced deployment:Unit attribute to align with other
deployment properties.