feat: management contract generation
Adds support for generating a management contract. The managment
contract uses Deployment Modes, Groups and Units to generate a document
outlining all deployment units that are enabled in a solution for a
provided deployment mode.

This contract can be used to sequence a deployment or template
generation process using a discovery based process.

The contract uses Stages to represent groups, with ordering based on the
membership definition of the deployment mode.

A step is created within each stage for deployment units that belong to
the group. They are ordered based on the deployment units priority.

Also introduces a new contract execution mode priority. Priority
execution mode allows for hybrid parallel and serial execution. Steps
must be executed from lowest (number ) priority to highest, and if two
steps have the same priority they can be executed in parallel.

The management contract replaces the existing unit list document that
only listed deployment units before