feat: dynamic model loading and scopes
Refactors models to be loading dyanmically based on command line
options.

Introduces the idea of scopes as part of models, scopes provide
different processing entrypoints to the data available within a model.

The majority of outputs generated throughout hamlet are based on the
Components scope which generates outputs in the scope of a component.
However if you want to generate outputs based on other data avaialble
within hamlet you had to do it as part of the component context or
override the entire processing layer

To support more freeform approaches to generating output a new scope,
views have been introduced. They allow for different outputs to be
cretaed with the view defining the complete processing process to find
the data required.