refactor: dynamic loading of subset marcos
Allows for component subset generation macros to be dynamically loaded
and to offer shared component macros, which can be used as a default
method for a specific subset macro.

Also allows for differnet subset macros based on the documentset that
has been requested. This allows for different outputs to be generated
based on the document set type under the Components scope

The refactor moves the test case and management contract output
generation under a shared component within the shared provider to align
with the existing component structure

BREAKING CHANGE: All Component macros now need to include an entrance as
part of their name to allow for different macros based on entrance