refactor: load all reference data from blueprint
In the setContext process we currently load each reference data type
based on when we require them, usually to assign it to a global
environment. Since the provider loading process defines all of the
possible referenceData sources available we can just load all of the
data as part of the bootstrap process. If specialised actions are
required then we can reload the data as required