feat: fragment migration to extensions
Introduces extensions which replace and extend the current fragment
implementation.

Extensions are self contained macros which can modify a components
context. The context is initially defined a component and outlines
configuration which can be modified or controlled by extensions. A
component invokes the extension processing once it has assembeled what
it needs for the context

once the processing is complete the context is then passed back to the
occurrence and it can proceed with building out the deployment

The migration also maintains support for the existing fragment process