refactor(attributesets): update schema ref path
When generating a JSONSchema file in hamlet, some schemas
need to refererence another one to provide their structure.
This is done through a "ref" attribute. Currently, there is
only one attributeset type - Links. Components that ref the
links schema need to be able to provide a path to that schema,
as it will be in another file. This commit changes that file
from the "metaparameter" file that was previously created to
use a new "attributeset" file which will be implemented shortly
in its place.