refactor(attributesets): meta schema to attrSet schema
During schema generation, a list of the other schemas that exist and
thus should be defined as references to it is provided. This commit
removes a manually maintained list of metaparameters to ref, and
replaces it with an array of the keys of all AttributeSets. This will be
less effort to maintain than a single list.