feature: composite objects to ref attributesets
As established in https://github.com/hamlet-io/engine/issues/1491,
the composite object structure should be able to define a reference to
an AttributeSet. This reference can in-engine be evaluated to construct
the final object, but enables attribute re-use.

This also aligns nicely with the JSONSchema "$ref" attribute, which
performs the same functionality within a schema.