fixup: eval references before attr processing
In the previous commit I established the processing behaviour for
references in a composite object structure. The problem with it was that
it was done during the looping over each attribute, and an attribute
could not update itself as the itterant.

This commit moves the handling of references prior to all other
attribute processing, and creates its own list of "evaluated"
attributes. Any references are then resolved and turned into attribute
children. The new "evaluated" array is then looped over to perform
existing attribute processing as expected.