Load plugins from CMDB (#1469)
* feat: provider loading definition in shared layers

Defines the configuration for loading providers through layers
Note these will only be supported in the shared provider layers as the
provider loading only runs within the context of the shared layer

* feat: add core provider support for layer loading

Use the active layers to determine the providers which can be loaded
Adds a metadata function which returns the loaded providers

* feat: add loader contract generation

Introduces a new entrance which generates a contract based output
detaling the requirements for an executor to load and configure
providers which have been configured in an entrance

The entrance has been called the loader as this could be easily extended
to load other sources such as a CMDB during the same processing pass

* feat: retrieve provider state from loader process

When providing a contract for the loader process we also support an
input which provides the state of providers which have been loaded. This
includes details such as git commits which we can then use for
versioning details

The initial implementation uses this state as metadata in contract
generation

* fix: minor fixes to support only loading shared

Some minor fixes required to support loading only the shared provider
which doesn't have context around regions or other details that we use
in setContext

* refactor: rename loading to plugins instead of providers

* refactor: move plugin state to its own option

* refactor: move seeding selection into bootstrap

* feat: plugin loading as part of boostrap process

* fix: missing value handling and metadata

* feat: add engine to metadata

* Update providers/shared/layers/layer.ftl

* refactor: naming and local hanlding

Updates naming to align with configuration convention for type specific
configuration options

Removes the path configuration for local paths and instead relies on the
plugin being available through the GENERATION_PLUGIN_DIRS environment
variable. This supports wider use and removes local directories from the
CMDB

* fix: add providers back on account properties

Co-authored-by: Ross Murray <ross.murray@gosource.com.au>