feat: MTA component
The "Message Transfer Agent" component represents the capability to
send or recieve emails. Each occurrence either supports sending or
receiving. If noth are required, two occurrences should be configured.

An MTA may have zero or more rule subcomponents. Rules are evaluated in
order and if they fire, execute a sequence of actions.

While debugging the MTA, a couple of logging improvements to the
component flow were also made.