fix: masterdata inclusion in blueprint
Ensure overrides in the blueprint of the masterdata are not lost
during assembly of the blueprintObject.

Previously, the merging of the masterdata with the blueprint caused
changes to be lost. This fix introduces a new function to permit the
blueprint to be rebased on the master data effectively ensuring that
blueprint data takes precedence over master data.

Some comments were also added to the bootstrap process that need
confirmation. If correct, then some of the bootstrap processing
(currently ocmmented out) can be removed.