Refactor - Rename Metaparameters to AttributeSets (#1498)
* feat(attributesets): rename metaparameter to attributesets

As covered in https://github.com/hamlet-io/engine/issues/1491
Metaparameters are to be renamed as AttributeSets. This better
aligns the intention of Metaparameters with reuseable and
reference-able sets of attributes.

* refactor: rename metaparameter file to attributeset

* refactor(attributesets): rename metaparam configuration object

As metaparameters are renamed to attributesets, the global
composite object structure must be renamed to match.

* refactor(attributesets): update links macro

* refactor(attributesets): update provider load of attrSets

* refactor(attributesets): update schema ref path

When generating a JSONSchema file in hamlet, some schemas
need to refererence another one to provide their structure.
This is done through a "ref" attribute. Currently, there is
only one attributeset type - Links. Components that ref the
links schema need to be able to provide a path to that schema,
as it will be in another file. This commit changes that file
from the "metaparameter" file that was previously created to
use a new "attributeset" file which will be implemented shortly
in its place.

* refactor(attributesets): meta schema to attrSet schema

During schema generation, a list of the other schemas that exist and
thus should be defined as references to it is provided. This commit
removes a manually maintained list of metaparameters to ref, and
replaces it with an array of the keys of all AttributeSets. This will be
less effort to maintain than a single list.

* refactor(attributesets): added attributeset schema

The attributeset schema replaces the metaparameter schema

* fixup(schemas): correct schema paths to match filenames

* fixup: lowercase atttributeset type

Co-authored-by: Ross Murray <rossmurray@protonmail.com>
Co-authored-by: Michael Foley <roleyfoley@gmail.com>