feat(registry): support for mutli region registry
Adds support for mutli region S3 based registries. When deploying
components like AWS lambda functions the zip artefact must be in the
same region as the lambda code. When deploying to multiple regions
inside of a solution we need to establish a central registry for the CI
process and then replicate the artefacts to the other regions.

Additonal region endpoints can be assigned through the RegionEndpoints
setting as part of the Registry setting definition.

To deploy the registy, the Account layer needs to be updated with the
replica region and the s3replica deployment unit deployed to secondary
regions