refactor: use name for url images
Instead of using the deployment unit for URL image source registry
references this PR changes to using the component name. With image url
components its possible to bypass our current image magement which
relies on the deployment unit. Updating the configuration allows for
multiple components with different images to be used in the same
deployment unit