fix: schema assignment of multiple data types
JSONSchema is able to define a property as "any of the following" data
types, defined as an Array of Objects, each object with a single
attribute - "type".

Currently the JSONSchema being generated will define this structure
incorrectly, placing each object inside of its own array. This is not
only incorrect JSONSchema but also invalid JSON more broadly.

This removes the additional arrays, so the resulting array isThis is not
only incorrect JSONSchema but also invalid JSON more broadly.

This commit removes the unnecessary level of arrays.

current structure:
[
  [
    {
      "type" : "string"
    }
  ],
  [
    {
      "type" : "bool"
    }
  ]
]

new structure:
[
  {
    "type" : "string"
  },
  {
    "type" : "bool"
  }
]