fix(schema): reference data as one schema per schema
Recently schema outputs were refactored to create
one output file per schema definition. This wasn't
behaving as expected for reference data, which was
creating individual output files, but each with all
schema defintions inside of it.

This commit corrects this and realigns the outputs
with that of Component schema type.