fix: flowlog configuration (#1531)
* fix: flowlog configuration

Remove/disable flag based configuration settings for flowlogs.

Support for flowlogs now needs to be explicitly included on the network
componenent. These changes ensure the default configuration does not
trigger the checks in place to report usage of the legacy format of
flowlog configuration.

* refactor: change attribute descriptions

Use the term `deprecated` to indicate the attributes should not be used
for new work.