Shared provider testing (#1532)
* feat: add shared testing provider

Adds a test provider for the shared provider which allows us to test
basic Occurrence structures as part of the shared provider

Since we are testing the externalservice we can also use this to test
our settings macros

* feat: add externalservice basic deployment

Adds basic deployment support for the external serivce. This essentially
creates a config file of the occurrence. It's useful for testing
external service development and also allows us to test the shared
provider with deployment tests

* feat: deployment test script for shared

Adds a testing script for the shared provider which performs some basic
health checks on the core engine focussed on generating components

* feat: jenkins testing based on test script

Uses the test script to deteremine units and create deployment tests
based on the shared provider

junit results are recorded as part of the testing

* fix: override engine location

* refactor: move externalservice deploy to test