feat: add raw ids for occurrence details
Adds the Raw Id field for Version and Instance. This is useful for user
facing output that reads these files. in the CMDB when we have "default"
as the instance or version Id its replaced with an empty string. This
makes naming and output in deployment easier. But harder to align the
whats in the CMDB with query commands