feat: validate composite object model
BREAKING CHANGE: The composite object model in Hamlet is used
throughout hamlet to define the various data types used by the engine.
These data types are then used within Hamlet Tenants/Accounts/Products
by end-users in order to configure their CMDB's.

Currently, to determine a given attribute's value the composite object
model simply says : here is the composite object, give me attribute
"x", normalise it, evalute its attriubute sets and return it to me.
The remainder of the composite object is not validated. In this way
it is trivial for a hamlet user to input a typo, an attribute in the
wrong object level or be otherwise wrong and not know it - hamlet
will work as expected so long as all the mandatory Attributes are
found.

This introduces a new parameter into the composite object model:
AllowAdditionalParameters.

False by default, this will provide fatal error and sufficient
debugging information in the event that the composite object
being evaluated contains configuration that is undefined.

Where necessary, this can then just be disabled on a per-attribute
basis.