fix: internaltest component definition (#1547)
* fix: undefined component type for internaltest

Moves the general definition of the internaltest component to the shared
provider. This allows for the component type to be loaded as part of the
shared provider. The state and setup definitions that we use for testing
are handled by the shared test provider

* refactor: move deployment group details to test

Moves as much as possible over to the sharedtest provider for internal
testing isolation