refactor: incorporate input system review feedback
- add validation of long form qualifiers (will need a further revision
once support for not allowing additional attributes is in place)
- add mechanism to define what filter attributes can be used during
long form qualification (think we should deprecate the short form)
- add entrance input seeding
- add initialisation of the new input system
- add input state change stack.

The stack will form the basis for switching inputs e.g. when following
links. The input state cache ensures the the overhead of recalculating
inputs only happens the first time a unique set of filter attributes is
encountered.

These changes are designed to introduce new functionality in parallel
with the existing process so should not break anything.