refactor: product layer attribute definitions
Adds in unnaccounted for but expected-to-be-there layer attribute
definitions.

The Product layer also provides the ability to define single-region
overrides for individual deployment-units. This attribute is
represented by a wildcard as the deployment-unit names are highly
specific.

Also removed the "frameworkObjectAttributes" from here, as they
are added to every object level in the blueprint outside of this
definition and do not need to be defined unless providing overrides.