feat: blueprint composite model
The composite object model for the blueprint has previously been
something that you had to about - no model for all the possible
attributes for the top-most level of configuration existed.

The cookiecutter templates would get you started with the necessary
layers, but only through experience (or reading the code) could
you know all possible options in the blueprint.

This introduces a new function to construct the model.