feat: new entrance - validate (the blueprint)
Introduces a new entrance which will perform validation
of a blueprint. The validation will notify you of any
additional attributes that are present where they should
not be.

This offers a "soft" validation mechanism whereby a user can ensure
that their configuration is valid, whilst not causing immediate
failure in existing configuration that may contain a typo or two.

Example usage:

hamlet entrance -i mock invoke-entrance -e validate -o .

This will use the mock input source along with the shared-provider
masterdata to form a blueprint, and validate it.