feat: Stop on fatal (#1556)
Add the ability to stop after a threshold number of fatal errors, as well
as the option for a fatal error to always cause processing to stop.

The default behaviour reflects current behaviour where there is no limit
so all fatal errors will be reported together.

When debugging complex behaviour, it is sometimes useful to know the
state of processing at a particular point, especially when later
processing throws a template error. This change permits a fatal error
with `stop=true` to be added to know the exact state of variables etc.

The threshold behaviour allows for different user preference e.g. stop
on first fatal error or permit multiple fatal errors to be recorded
before processing stops.