Feat: New Entrance - validate (the blueprint) (#1548)
* feat: validate composite object model

BREAKING CHANGE: The composite object model in Hamlet is used
throughout hamlet to define the various data types used by the engine.
These data types are then used within Hamlet Tenants/Accounts/Products
by end-users in order to configure their CMDB's.

Currently, to determine a given attribute's value the composite object
model simply says : here is the composite object, give me attribute
"x", normalise it, evalute its attriubute sets and return it to me.
The remainder of the composite object is not validated. In this way
it is trivial for a hamlet user to input a typo, an attribute in the
wrong object level or be otherwise wrong and not know it - hamlet
will work as expected so long as all the mandatory Attributes are
found.

This introduces a new parameter into the composite object model:
AllowAdditionalParameters.

False by default, this will provide fatal error and sufficient
debugging information in the event that the composite object
being evaluated contains configuration that is undefined.

Where necessary, this can then just be disabled on a per-attribute
basis.

* fix: remove invalid configuration from masterdata

* fix: define Region attr on all layers

* fixup: additional commonParams

* fix: define Domain on account/product layers

* fix: add missing product layer build child attribute

* fixup: rework the logic of the attribute validation

* fixup: check for subobjects of both variants

* refactor: update coreComponent config with missing attr defs

* refactor: frameworkObjectAttributes fleshed out into attribute defs

* fix: update attribute SubObjects param across engine

* fix: missing syntax

* refactor: add extension definition as versions child

* refactor: product layer attribute definitions

Adds in unnaccounted for but expected-to-be-there layer attribute
definitions.

The Product layer also provides the ability to define single-region
overrides for individual deployment-units. This attribute is
represented by a wildcard as the deployment-unit names are highly
specific.

Also removed the "frameworkObjectAttributes" from here, as they
are added to every object level in the blueprint outside of this
definition and do not need to be defined unless providing overrides.

* fixup: remove unused attribute property

* fix: more missing syntax

* fixup: framework object attributes

the new structure for the framework object attributes
(now using the composite object model attribute structure
instead of strings) receives an inclusion of "Enabled".

Also removes the definition of the Qualifiers structure completely, as
there are several valid sub-structures.

* feat: blueprint composite model

The composite object model for the blueprint has previously been
something that you had to about - no model for all the possible
attributes for the top-most level of configuration existed.

The cookiecutter templates would get you started with the necessary
layers, but only through experience (or reading the code) could
you know all possible options in the blueprint.

This introduces a new function to construct the model.

* feat: get blueprint object helper function

getBluePrintObject() provides a wrapper around getCompositeObject().

It walks the attributes (and their children) adding the
"Framework Object Attributes" to every level. In this way these
attributes are available to be defined at every level of the blueprint.

* fixup: perform object validation at topmost-level

previously was only validating children

* feat: new entrance - validate (the blueprint)

Introduces a new entrance which will perform validation
of a blueprint. The validation will notify you of any
additional attributes that are present where they should
not be.

This offers a "soft" validation mechanism whereby a user can ensure
that their configuration is valid, whilst not causing immediate
failure in existing configuration that may contain a typo or two.

Example usage:

hamlet entrance -i mock invoke-entrance -e validate -o .

This will use the mock input source along with the shared-provider
masterdata to form a blueprint, and validate it.

* fixup: remove deployment:Priority/Group from core config

* fixup: dont require extensions on Versions attr

* fixup rename function to getBlueprintConfiguration

* fixup: add all resourcegroup attrs to valid config for validation

* fixup: remove erroroneous settingschildren

* fixup: define domainChildConfiguration and refactor existing uses

* fixup: create validate view and move out of blueprint view

Co-authored-by: Ross Murray <rossmurray@protonmail.com>