fix: hanlding of log level lookup
The log lookup process uses a list to map the string log level to a log
level number. Some entries in the list are empty to handle extra level
spaces.

When the string level is converted it used a startswith check which
considers an empty string as true rather than false

This PR ensures that the log level uses an explicit equals and also
checks that the item in the level list has content