fix: disable qualifier transformer
At present, solutions have been designed on the assumption that layer
ids could be used as the basis for the qualification. The qualifier
transformer currently only works on the layer names.

Once layers are better supported in the input processing, the transformer
can be updated to include layer ids in the filter and the transformer
re-enabled.