refactor: state processing
Switch lookup of state to use the state pipeline.

- core output lookup now supported via state pipeline
- cloud provider plugins need transformers to normalise their stack
outputs to the format expected by the shared cmdb state seeder
- support fixtures to allow all outputs to be validate
- support simulation of outputs that aren't present in the cmdb
- outputs validated during normalisation processing

In order to be able to validate stacks during normalisation, the
implementation of validation in `getCompositeObject()` and the `validate`
entrance had to be modified to avoid state access within a transformer.

A refinement was also made when outputting logs where the depth of the
objects in the log can be controlled. This is useful where large objects,
such as masterdata, are included in the output. By default, messages
are logged to their full depth for backwards compatability.