refactor: state processing (#1605)
* refactor: state processing

Switch lookup of state to use the state pipeline.

- core output lookup now supported via state pipeline
- cloud provider plugins need transformers to normalise their stack
outputs to the format expected by the shared cmdb state seeder
- support fixtures to allow all outputs to be validate
- support simulation of outputs that aren't present in the cmdb
- outputs validated during normalisation processing

In order to be able to validate stacks during normalisation, the
implementation of validation in `getCompositeObject()` and the `validate`
entrance had to be modified to avoid state access within a transformer.

A refinement was also made when outputting logs where the depth of the
objects in the log can be controlled. This is useful where large objects,
such as masterdata, are included in the output. By default, messages
are logged to their full depth for backwards compatability.

* refactor: use input pipeline for all seeding

- use the input pipelines to control the population of all input data
- add layer, plugin and module stages
- add normalisation of inputs to implement the ordering of masterdata,
fixtures, cmdb and module data
- input pipeline processing now supports restarts to allow plugin and
provider loading
- remove files associated with previous input seeding mechanism and
convert any fixture data to config seeders