feat: add prefixing function for attributes
Adds a prefixing function to append a provider specific prefix to Names,
Values and Defaults on attributes. This allows for providers to add
their own attributes into configuration namespaces and show that it is
independent from the shared configuration.

If shared: is appended to the value that is being assesed the provider
prefix will be ommitted