refactor: create compute instance config set
Moves the image configuration under a general compute instance section
This section is intended to define the image and boostrapping
configuration of compute instances used in the component.

The Operating System configuration option is available for bootstrap
configuration so that conditional logic can be used based on OS

OSPatching has moved under this configuration section.