feat: add attirbutes during extension processing (#1625)
* feat: add prefixing function for attributes

Adds a prefixing function to append a provider specific prefix to Names,
Values and Defaults on attributes. This allows for providers to add
their own attributes into configuration namespaces and show that it is
independent from the shared configuration.

If shared: is appended to the value that is being assesed the provider
prefix will be ommitted

* refactor: add base condition to prefix attributes

refactor providerPrefix into a  a generic function which can be used
to append different parts of attributes

Names - Append a preffix to each name
Values - Append a prefix to each value in the list
Default - Append a prefix to a string based default value

The level of prefixing can be controlled based on the level of chidren
we are at and prefixing is disabled if the disablePrefix is already
present on the value to be prefixed

add support for adding prefixed attributes when setting provider
attribute extensions. If a attribute with the same name can't be found
in the shared attributes the extra attributes are added and prefixed
with the provider that has added then