feat: computeTask configuration extension macro
Adds the computeTaskConfigSection macro which updates the context of an
extension with the configuration required to complete ComputeTasks.

The content can be in the format that the virtual machine requires to
perform the configuration task

The computeTaskType should include one or many compute tasktype ids that
the content of the configuration completes

The priority with lowest first, provides a way for configuration tools
to order different sections of the content as appropriate.

The engine specifies the configuration tool that the content is for to
allow for multiple tools to be used at once