refactor: align bootstrap naming with computetask
Updates the bootstrap confituration with the new definition of compute
tasks. The boostrap options were added but not implemented so this
should have no impact on current components

UserTasksRequired allows for users to include required tasks that must
be included across all deployed components. This is useful for policy
enforcement and user defined configuration

Sets a default value for the ec2 Image source