feat: add attribute extension support
Adds support for creating an extended attribute set. This works along
the same line as attribtue extensions but with attribute sets.

When added an extended attribute set you provide a base type and a
provider. The same extension processing is then used to generate the new
attribute set.

This creates a new AttributeSet and there isn't a link between the base
attribute set and the new extended one