feat: handle log based exceptions in engine
Adds support for handling exceptions generated through log messages in
the engine itself. This allows for the engine to stop processing if it
finds logs which match the stop level

Adds support for in-line log messages in the engine itself. Once an
output has been generated the content of the output is scanned for the
presene of Hamlet<Severity>: messages which will be included in the
logMessages found during processsing

Updates the stop message action to use the logWriter process to output
the context of the stop. This allows for extra logs to be included as
required.