feat: add healthcheck component definition
Adds the health check component which is intended to implement the
function of transaction monitoring on deployed components.

Two types of health checks have been defined, simple and complex.

Simple performs basic probe based checks against a provided
destination and uses the port configuration to define basic probes

Complex is a scripted health check intended to be run in a serverless
compute environment and is capable of performing complex healthchecks
with full synthetic transactions