feat(ecs): add compute provider support on service
Adds support for specifying the compute provider on the ecs service.
When using compute providers the providers available depends on the
engine. So we set the default provider to the provider recommended by
the engine

Removes vendor specific attributes from the container definitions and
instead uses attribute extensions to define them in the provider