feat: make refresh input state routine public
Expose a method for general use to request a refresh of the input state.

This is used internally in the input source whenever the seeder
configurtion of the current input source is changed.

The new use case is for multi-pass processing where the configuration
of seeders isn't changing, but the data provided by the seeder varies
per pass, so it needs to request the input to be reassessed on each
pass.

Also remove technical debt associated with the introduction of the
input pipelines
- remove the global `commandLineOptions` variable and associated support
as all command line option access should be via the input state
accessor method
- remove the masterdata accessor routine as all masterdata management
is now done via seeders in the config input pipline.