feat: input qualification (#1695)
* feat: input qualification

Re-enable qualification

Fixes #1599

Reverses PR #1600

For backwards compatability, ensure input qualification checks for
ids and names of layers. Best practice should be to encourage the use of
names and long form qualification but a number of existing cmdbs use
ids and short form qualifiers.

This PR also adds some longer form alternatives for the `info` and `warn`
macros, as well as an alternative `fatal` macro which always stops
processing.

A followup PR will remove the use of specific checks for qualification
via the segmentQualifiers global variable.

* refactor: Capture qualification outcomes

Add an annotated copy of the state to the inputState showing which
qualifiers were or weren't applied.

This is intended to form the basis of tooling to assist users determine
the effect of the provided qualifiers where the results produced are
unexpected.