refactor: remove direct references to region
There is a mix of use of `region` and `regionId` reflecting the fact
that the composite assembly process initialises `region` while
`getContext` populates regionId.

With dynamic loading, references to region need to be made more dynamic
so removal of direct references to the input value is the first step in
preparing for this.