refactor: remove context flow
The context flow was an experiment aimed at delivering an amalgam of
ideas that have evolved to become input source qualification and the
dynamic loading of cmdbs.

It makes sense to remove it to avoid confusion for anyone trying to
understand the engine plugin, and also given the implementation wasn't
complete and has never been used.